added support for actions
actions are only called when there is a change of availability
(a service comes up on a new host, or an existing service goes down)
of any service that a host subscribes to.

update_hosts is now no longer called by default, but rather is provided
as an action which can be subscribed to. Subscribe to this action by
specifying

(in ec2 user-data):
action=UpdateHosts.new(local_ipv4, local_name, public_ipv4, public_name)

(on the command line, not on ec2):
-A "UpdateHosts.new('127.0.0.1', 'local_name', '<public ip>', 'public_name')"

Since the action string is eval'd inside ec2-discovery.rb, putting calls like
public_ipv4 in the first example is fine (but relies on ec2)